pyproject.toml
[project]
name = "terminoloogia"
version = "0.1.0"
description = "EN ↔ ET andmevaldkonna mõistete sõnastik"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
  "mkdocs>=1.6.1",
  "mkdocs-material>=9.6.14",
]
harvest = [
  "requests>=2.31",
  "lxml>=5.0",
  "spacy>=3.7",
]
